Add endpoint option
