+++ src/index.ts
@@ -1,2 +1,2 @@
-export {AuthsignalBrowser} from "./AuthsignalBrowser";
+export {AuthsignalBrowser} from "./authsignal-browser";
 export * from "./types";
