+++ src/popup-handler.ts
