+++ src/types.ts
@@ -1,9 +1,4 @@
-export type ChallengeInput = {
-  challengeUrl: string;
-  mode?: "popup" | "redirect";
-};
-
-export type MfaInput = {
+export type HandleUrlInput = {
   url: string;
   mode?: "popup" | "redirect";
 };
@@ -27,4 +22,13 @@ export type AuthsignalOptions = {
    * Name of id cookie. __eventn_id by default
    */
   cookieName?: string;
+
+  /**
+   * A URL pointing to the Authsignal MFA page.
+   */
+  endpoint?: string;
 };
+
+export enum AuthsignalWindowMessage {
+  AUTHSIGNAL_CLOSE_POPUP = "AUTHSIGNAL_CLOSE_POPUP",
+}
